Polish AB1 and Fixed Elle's Stats
Added a bit more text and flair to AB1 including recuritment theme, and talks between Adam, Elle, and Zane. Changed some AI a bit and gave text to brigands. Kajarn also has some text now and reinforcments towards the end of the chapter to not make the end feel so empty. Removed the shields on some enemies since it made them way to bulky.